requirements.txt
PyMuPDF>=1.24.0

openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0

Pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.0.0